Move isHiddenReasoningModel
